requirements.txt
fastapi
uvicorn[standard]
motor
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
email-validator
